+++ controllers/auth/index.js
@@ -1,9 +1,15 @@
-const {
+const { signup } = require("./signup");
+const { login } = require("./login");
+const { getCurrent } = require("./getCurrent");
+const { logout } = require("./logout");
+const { updateSubscription } = require("./updateSubscription");
+const { updateAvatar } = require("./updateAvatar");
+
+module.exports = {
   signup,
   login,
   getCurrent,
   logout,
   updateSubscription,
-} = require("./auth");
-
-module.exports = { signup, login, getCurrent, logout, updateSubscription };
+  updateAvatar,
+};
