+++ controllers/auth/login.js
@@ -1,0 +1,26 @@
+const { User } = require("../../models");
+const { Unauthorized } = require("http-errors");
+const jwt = require("jsonwebtoken");
+const { SECRET_KEY } = process.env;
+
+const login = async (req, res) => {
+  const { email, password, subscription } = req.body;
+  const user = await User.findOne({ email });
+  if (!user || !user.comparePassword(password)) {
+    throw new Unauthorized(`Email or password is wrong`);
+  }
+  const payload = { id: user._id };
+  const token = jwt.sign(payload, SECRET_KEY, { expiresIn: "10h" });
+  await User.findByIdAndUpdate(user._id, { token });
+  res.status(200).json({
+    data: {
+      user: {
+        email,
+        subscription,
+        token,
+      },
+    },
+  });
+};
+
+module.exports = { login };
