+++ middlewares/index.js
@@ -2,5 +2,5 @@ const { validation } = require("./validation");
 const { auth } = require("./auth");
 const { asyncWrapper } = require("./wrappers");
 const { isValidId } = require("./isValidId");
-
-module.exports = { validation, auth, asyncWrapper, isValidId };
+const { upload } = require("./upload");
+module.exports = { validation, auth, asyncWrapper, isValidId, upload };
