+++ middlewares/upload.js
@@ -1,0 +1,22 @@
+const multer = require("multer");
+const path = require("path");
+
+const tempDir = path.join(__dirname, "../", "tmp");
+
+const multerConfig = multer.diskStorage({
+  destination: (req, file, cb) => {
+    cb(null, tempDir);
+  },
+  filename: (req, file, cb) => {
+    cb(null, file.originalname);
+  },
+  limits: {
+    fileSize: 2048,
+  },
+});
+
+const upload = multer({
+  storage: multerConfig,
+});
+
+module.exports = { upload };
