+++ package.json
@@ -14,11 +14,14 @@
     "cross-env": "7.0.3",
     "dotenv": "^16.0.3",
     "express": "4.17.1",
+    "gravatar": "^1.8.2",
     "http-errors": "^2.0.0",
+    "jimp": "^0.22.4",
     "joi": "^17.7.0",
     "jsonwebtoken": "^9.0.0",
     "mongoose": "^6.8.4",
     "morgan": "1.10.0",
+    "multer": "^1.4.5-lts.1",
     "uuidv4": "^6.2.13"
   },
   "devDependencies": {
